Renamed vecmem::reverse_iterator to vecmem::details::reverse_iterator.
Also moved it under the containers/ sub-directory, since it will really
only ever be needed by the container types.

This is in preparation for the introduction of the jagged vector
specific iterator.
