+++ core/CMakeLists.txt
@@ -32,6 +32,9 @@ vecmem_add_library( vecmem_core core SHARED
    "include/vecmem/containers/impl/vector_buffer.ipp"
    "include/vecmem/containers/data/vector_view.hpp"
    "include/vecmem/containers/impl/vector_view.ipp"
+   # Iterator types.
+   "include/vecmem/containers/details/reverse_iterator.hpp"
+   "include/vecmem/containers/impl/reverse_iterator.ipp"
    # Allocator
    "include/vecmem/memory/allocator.hpp"
    "include/vecmem/memory/allocator.ipp"
@@ -48,7 +51,5 @@ vecmem_add_library( vecmem_core core SHARED
    "src/memory/contiguous_memory_resource.cpp"
    "include/vecmem/memory/contiguous_memory_resource.hpp"
    # Utilities.
-   "include/vecmem/utils/reverse_iterator.hpp"
-   "include/vecmem/utils/reverse_iterator.ipp"
    "include/vecmem/utils/type_traits.hpp"
    "include/vecmem/utils/types.hpp" )
