+++ core/include/vecmem/containers/array.hpp
@@ -8,8 +8,8 @@
 
 // Local include(s).
 #include "vecmem/containers/data/vector_view.hpp"
+#include "vecmem/containers/details/reverse_iterator.hpp"
 #include "vecmem/memory/memory_resource.hpp"
-#include "vecmem/utils/reverse_iterator.hpp"
 #include "vecmem/utils/types.hpp"
 
 // System include(s).
@@ -67,9 +67,10 @@ namespace vecmem {
       typedef const_pointer      const_iterator;
 
       /// Reverse iterator type
-      typedef vecmem::reverse_iterator< iterator > reverse_iterator;
+      typedef vecmem::details::reverse_iterator< iterator > reverse_iterator;
       /// Constant reverse iterator type
-      typedef vecmem::reverse_iterator< const_iterator > const_reverse_iterator;
+      typedef vecmem::details::reverse_iterator< const_iterator >
+         const_reverse_iterator;
 
       /// @}
